+++ Resources/Locale/en-US/deltav/roboisseur/roboisseur.ftl
@@ -8,7 +8,7 @@ roboisseur-request-impatient-1 = There is no time left! Bring me {$item}, immedi
 roboisseur-request-impatient-2 = My market opportunity is dwindling! Quickly, you fools, {$item}!
 roboisseur-request-impatient-3 = Damn you all! Deliver me a {$item} already!
 roboisseur-request-second-1 = I humbly request one {$item} for one of my snowbound clients.
-roboisseur-request-second-2 = The frozen planet in this sector requests an {$item}!
+roboisseur-request-second-2 = The frozen planet in this sector requests a {$item}!
 roboisseur-request-second-3 = A local ice cream vendor is looking to trade for a {$item}.
 roboisseur-thanks-1 = Mmm, I'm sure my client will be most pleased. 
 roboisseur-thanks-2 = I cannot believe you were able to source it, but I suppose you are the best. 
@@ -16,7 +16,7 @@ roboisseur-thanks-3 = When I first heard there were such skilled creatives on th
 roboisseur-thanks-4 = Unlike... most of your coworkers, my clients can afford these delicacies!
 roboisseur-thanks-5 = You have my word, my nominal fees are... more than reasonable. 
 roboisseur-thanks-second-1 = You shared your hopsitality, now they share their frozen delights.
-roboisseur-thanks-second-2 = A cold treat should be reward enough, but have some more too.
+roboisseur-thanks-second-2 = A cold, sweet treat should be reward enough, but have some cold, hard cash too.
 roboisseur-thanks-second-3 = While the iced delicacy alone is currency to them, they understand your needs.
 roboisseur-thanks-second-4 = Summer on their planet lasts for only two weeks. You have their thanks.
 roboisseur-thanks-second-5 = My client's pockets are as deep as their world is cold.
