+++ Resources/Prototypes/DeltaV/NPC/roboisseur.yml
@@ -2,12 +2,12 @@
   parent: BaseStructure
   id: Roboisseur
   name: Mr. Butlertron
-  description: It asks for food to deliver to exotic customers across the cosmos. Maybe it's got a Psionic teleporter?
+  description: It asks for food to deliver to exotic customers across the cosmos. Powered by the latest technology in bluespace food delivery.
   components:
   - type: Sprite
     noRot: true
     drawdepth: Mobs
-    offset: 0.0, 0.5
+    offset: 0.0, 0.0
     sprite: DeltaV/Structures/Machines/roboisseur.rsi
     layers:
     - state: roboisseur-1
